Add render thread support to Editor
Replaces DllImport with LibraryImport for engine interop, adds new enums for virtual keys, introduces RenderThread and FrameInfo for multi-surface rendering, and refactors RenderSurfaceHost for improved resource management. Updates EngineAPI and related classes for safer resource handling and component updates, and modernizes array initializations. Also fixes minor code style issues and improves exception handling.

+++ Editor/Content/UploadedAsset.cs
@@ -115,7 +115,7 @@ public class UploadedAsset
                 {
                     AssetInfo = assetInfo,
                     Metadata = asset.GetMetadata(),
-                    ContentId = EngineAPI.CreateResource(data, assetInfo.Type),
+                    ContentId = EngineAPI.CreateResource(data, (int)assetInfo.Type),
                     ReferenceCount = 1,
                     _referencedAssets = referencedAssets
                 };
